Add file descriptor persistence
1. Provides signigificant performance improvement in large OVSDB
installs such as southdb in OVN by removing the current "feature" of
"trying to break an EAGAIN wall" in all streams. All streams will
attempt IO IF AND ONLY IF there is IO pending. If there is no IO
signalled by (e)poll they will not attempt any send/recv. This saves
600 microseconds on each iteration in json rpc server for 2000 sessions
(measured).

2. Allows support for persistent in-kernel frameworks which replace
poll on most OSes. These frameworks imply and require a persistent
fd design.

3. Adds support for EPOLL on Linux.

4. Fixes the logic in the current poll loop which claims to be "additive"
on input and output events, but in fact is not - they are wiped clean
every time and reinstated by upper layers after that.

5. Essential first step in setting up the IO loop to allow multi-threading
in SSL/JSON/IO processing (needed for northd and ovsdb in OVN) at a later
date.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>