Fix -EINTR handling
There is no need to perform "catch signal, then pipe a notification"
equilibristics on any unix at present. In fact, there has been no
need since pre-NPTL glibc and Linux kernel 2.4. If there is a Unix
derivative where this is broken it has to walk its own specific
and not destroy the performance for everyone else.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>