Prevent excessive contention of the logging mutex
The log levels for a module which specify "should I log or should
I go" can and should be checked witout a mutex.

The only variable which needs to be protected by a mutex is the fd
and there is no point checking for the fd being valid if all
loglevel parameters say "do not log".

This is an obvious and quite unnecessary performance bottleneck.
Instead of doing something productive, any thread that runs across
a VLOG statement at any level (2300+ times in the code) will
instead of that try to contend the logging mutex.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>