Errata - Revert the Potemkin Village in backlog reporting
This reverts a previous attempt to fix the issue caused by
presence/absense of output being incorrectly equated to
presense/absense of backlog in jsonrpc_send_block()

Backlog is now reported everywhere and not hidden for async.

This may result in some extra CPU cycles when servicing the
unixctl requests to ovsdb, but should otherwise have no
effect elsewhere.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>