Make waiting of POLLOUT explicit.
I am surprised (and I will investigate) that this is not
triggered in the normal (not async io) case.

It is definitely present in the async case - the result is
output being stuck. However, because when there is pending
output backlog input is stuck too, the end result is a
deadlock - the RPC dies.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>