Fix possible stall when input buffer fills up
If the input buffer has filled up and the main ovsdb
processing thread has failed to keep up we stop listening
to POLLINs from the file descriptor. That may result in a
hang with a connection drop an reconnect.

This adds a "kick" - if json rpc notices that it is working
off a full buffer it kicks the IO thread after emptying it.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>