ovsdb: Run monitor processing in parallel
Monitor processing performs a full walk of all rows in all
tables referenced in a monitor.

The rows are internally represented as a hash map.

This operation can be run in parallel where thread M out N is
walking hash buckets M, N+M, etc.

Running inter-thread IPC for only a handful of datum values
is not likely to be efficient, hence there is a cut-off. Tables
smaller than the cut-off value are run in the main thread.
Larger tables are run via multiple threads.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>